refactor: change schema filenames, fix import statements
