+++ backend/schemas/rooms.py
